Update [slug].js
showResults shows the new results component on button click. showButton to hide the button if something is wrong with the responseObject so we don't show an empty page. responseObject to save users inputs as they move through the calculator. Is then sent to the Results component to be shown with the questions